musl: fileio: check if file stream is writable before writing
Even the file is opened as the read-only mode, fputs() and fputc()
by musl succeed but nothing actually written. Hence, without this
change, write_string_stream() will be succeed even the file stream is
not writable.

+++ src/basic/hostname-util.h
@@ -5,7 +5,7 @@
 #include "strv.h"
 
 /* HOST_NAME_MAX should be 64 on linux, but musl uses the one by POSIX (255). */
-#define LINUX_HOST_NAME_MAX CONST_MIN((size_t) HOST_NAME_MAX, 64U)
+#define LINUX_HOST_NAME_MAX CONST_MIN((size_t) HOST_NAME_MAX, (size_t) 64)
 
 char* get_default_hostname_raw(void);
